rename timer module
